pyproject.toml
[tool.poetry]
name = "fincal"
version = "0.1.0"
description = ""
authors = ["Eric <eric@eruditis.com>"]
readme = "README.md"
packages = [{include = "fincal"}]

[tool.poetry.dependencies]
marketholidays = ">=0.11"
pandas = "*"
python = ">=3.10,<3.11.4"
python-dateutil = "^2.8.2"
pytz = "^2022.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
